Bug 2029889 - Cherry-pick changes from https://github.com/harfbuzz/harfbuzz/pull/5915 r=gfx-reviewers,lsalzman
Differential Revision: https://phabricator.services.mozilla.com/D294039

UltraBlame original commit: 603220a38119c844c8a0f97b14ee169fc012273b

+++ gfx/harfbuzz/src/OT/Layout/GPOS/PairSet.hh
@@ -24,11 +24,11 @@ struct PairSet : ValueBase
   public:
   DEFINE_SIZE_MIN (2);
 
-  static unsigned get_size (unsigned len1, unsigned len2)
+  static size_t get_size (unsigned len1, unsigned len2)
   {
     return Types::HBGlyphID::static_size + Value::static_size * (len1 + len2);
   }
-  static unsigned get_size (const ValueFormat valueFormats[2])
+  static size_t get_size (const ValueFormat valueFormats[2])
   {
     unsigned len1 = valueFormats[0].get_len ();
     unsigned len2 = valueFormats[1].get_len ();
@@ -39,7 +39,7 @@ struct PairSet : ValueBase
   {
     const ValueFormat *valueFormats;
     unsigned int len1; 
-    unsigned int stride; 
+    size_t stride; 
   };
 
   bool sanitize (hb_sanitize_context_t *c, const sanitize_closure_t *closure) const
